Move pagination under components-folder
